# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Jaseo Job Tracker
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pm/proposals:
    get:
      operationId: listPmProposals
      tags: [pm-proposals]
      summary: Browse proposals (public)
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["integer", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PmProposal"
    post:
      operationId: createPmProposal
      tags: [pm-proposals]
      summary: Create a new proposal (seller only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalInput"
      responses:
        "201":
          description: Created proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PmProposal"
        "403":
          description: Forbidden
  /pm/proposals/my:
    get:
      operationId: listMyProposals
      tags: [pm-proposals]
      summary: List seller's own proposals
      responses:
        "200":
          description: Seller's proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PmProposal"
        "403":
          description: Forbidden
  /pm/proposals/{id}:
    get:
      operationId: getPmProposal
      tags: [pm-proposals]
      summary: Get proposal detail (public)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PmProposal"
        "404":
          description: Not found
    patch:
      operationId: updatePmProposal
      tags: [pm-proposals]
      summary: Update a proposal (seller, own proposals only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalUpdate"
      responses:
        "200":
          description: Updated proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PmProposal"
        "403":
          description: Forbidden
  /pm/proposals/{id}/moderate:
    patch:
      operationId: moderatePmProposal
      tags: [pm-admin]
      summary: Approve or reject a proposal (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalModeration"
      responses:
        "200":
          description: Moderated proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PmProposal"
        "403":
          description: Forbidden
...
components:
  schemas:
...
    PmProposal:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        price:
          type: integer
        fileKey:
          type: ["string", "null"]
        previewImageKey:
          type: ["string", "null"]
        previewImageKey2:
          type: ["string", "null"]
        previewImageKey3:
          type: ["string", "null"]
        status:
          type: string
        totalSales:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - sellerId
        - title
        - description
        - price
        - status
        - totalSales
        - createdAt
    ProposalInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        categoryId:
          type: ["integer", "null"]
        price:
          type: integer
        fileKey:
          type: ["string", "null"]
        previewImageKey:
          type: ["string", "null"]
        previewImageKey2:
          type: ["string", "null"]
        previewImageKey3:
          type: ["string", "null"]
      required:
        - title
        - description
        - price
    ProposalUpdate:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        price:
          type: ["integer", "null"]
        fileKey:
          type: ["string", "null"]
        previewImageKey:
          type: ["string", "null"]
        previewImageKey2:
          type: ["string", "null"]
        previewImageKey3:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
    ProposalModeration:
      type: object
      properties:
        status:
          type: string
        adminNote:
          type: ["string", "null"]
      required:
        - status
...